Create TRUCK BOOKING MOBILE APPLICATION – VLIFT
In the realm of logistics, change is on the horizon as we eagerly anticipate the launch of an innovative mobile truck booking application in January 2024. Geared towards male drivers, this app aims not only to streamline the booking process but also to revolutionize the trucking experience in the bustling metropolis of Delhi, India.